refactor: rename NoAgent to UnknownAgent, make publicly usable
