+++ src/EntityDb.Common/Agents/UnknownAgentSignature.cs
@@ -1,0 +1,6 @@
+﻿namespace EntityDb.Common.Agents;
+
+/// <summary>
+///     Represents the signature of an unknown actor.
+/// </summary>
+public record UnknownAgentSignature;
