--- test/EntityDb.Common.Tests/Implementations/Agents/NoAgentAccessor.cs
@@ -1,11 +1,0 @@
-﻿using EntityDb.Abstractions.Agents;
-
-namespace EntityDb.Common.Tests.Implementations.Agents;
-
-public class NoAgentAccessor : IAgentAccessor
-{
-    public IAgent GetAgent()
-    {
-        return new NoAgent();
-    }
-}
