--- test/EntityDb.Common.Tests/Implementations/Agents/NoAgentSignature.cs
@@ -1,3 +1,0 @@
-﻿namespace EntityDb.Common.Tests.Implementations.Agents;
-
-public record NoAgentSignature;
